+++ tests/integration/test_integration_json.py
@@ -6,16 +6,14 @@ import os
 import unittest
 from datetime import datetime
 
-from amazonorders.orders import AmazonOrders
-from amazonorders.session import AmazonSession
-from tests.testcase import TestCase
+from tests.integrationtestcase import IntegrationTestCase
 
 PRIVATE_RESOURCES_DIR = os.path.normpath(
     os.path.join(os.path.abspath(os.path.dirname(__file__)),
                  "private-resources"))
 
 
-class TestIntegrationJSON(TestCase):
+class TestIntegrationJSON(IntegrationTestCase):
     """
     The two JSON files committed to "private-resources" are provided as examples of the syntax. Any other
     files created in "private-resources" will be ignored by ``.gitignore``.
@@ -76,34 +74,13 @@ class TestIntegrationJSON(TestCase):
     With this syntax, multiple ``Orders`` from the response can be asserted against. The indexed dictionaries under
     the ``orders`` key then match the assertion functionality when testing against a single order, meaning you
     define here the fields and values under the ``Order`` that you want to assert on.
-
-
     """
-    amazon_session = None
 
     def __init__(self, method_name, filename=None, data=None):
         super(TestIntegrationJSON, self).__init__(method_name)
 
         self.filename = filename
         self.data = data
-
-    @classmethod
-    def setUpClass(cls):
-        cls.credentials_found = os.environ.get(
-            "AMAZON_USERNAME") and os.environ.get("AMAZON_PASSWORD")
-
-        cls.amazon_session = AmazonSession(os.environ.get("AMAZON_USERNAME"),
-                                           os.environ.get("AMAZON_PASSWORD"))
-        cls.amazon_session.login()
-
-        cls.amazon_orders = AmazonOrders(cls.amazon_session)
-
-    def setUp(self):
-        if not self.credentials_found:
-            self.fail(
-                "AMAZON_USERNAME and AMAZON_PASSWORD environment variables not set")
-
-        self.assertTrue(self.amazon_session.is_authenticated)
 
     def run_json_test(self):
         print(f"Info: Dynamic test is running from JSON file {self.filename}")