+++ amazonorders/constants.py
@@ -59,6 +59,7 @@ BASE_HEADERS = {
 SIGN_IN_FORM_SELECTOR = "form[name='signIn']"
 MFA_DEVICE_SELECT_FORM_SELECTOR = "form#auth-select-device-form"
 MFA_DEVICE_SELECT_INPUT_SELECTOR = "input[name='otpDeviceContext']"
+MFA_DEVICE_SELECT_INPUT_SELECTOR_VALUE = "value"
 MFA_FORM_SELECTOR = "form#auth-mfa-form"
 CAPTCHA_1_FORM_SELECTOR = "form.cvf-widget-form-captcha"
 CAPTCHA_2_FORM_SELECTOR = "form:has(input[id^='captchacharacters'])"