+++ CHANGELOG.md
@@ -3,9 +3,9 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/alexdlaird/amazon-orders/compare/1.1.0...HEAD)
+## [Unreleased](https://github.com/alexdlaird/amazon-orders/compare/1.1.1...HEAD)
 
-## [1.1.0](https://github.com/alexdlaird/amazon-orders/compare/1.0.16...1.1.0) - 2024-04-09
+## [1.1.1](https://github.com/alexdlaird/amazon-orders/compare/1.0.16...1.1.1) - 2024-04-09
 ### Added
 - Build improvements.
 - Test improvements.